注销（个人中心页面、管理员页面、审核者页面）

+++ src/main/resources/static/AdminPage/js/index.js
@@ -81,8 +81,22 @@ function judgeLogin() {
         }
     })
 }
-function logoff(){
-    alert(1);
-    sessionStorage.clear();   //清除所有session值
-    window.location.href='/Home/login.html';
+//注销
+function cancel() {
+    $.ajax({
+        url:"/user/cancel",
+        type:"get",
+        success:function (resultData) {
+            resultData = JSON.parse(resultData);
+            if (resultData.meta.status === "200") {
+                sessionStorage.clear();   //清除所有session值
+                window.location.href='/Home/login.html';
+            }else{
+                layer.msg('操作失败!', {
+                    icon: 5, //红色不开心
+                });
+            }
+        }
+    })
+
 }
